Refactor the state machine into more distinct parts
State machine now only handles the actual transitions and the related
logic, but it is driven by a separate "Program" that receives the state
machine as a parameter. State machine is thus independent on how it is
being driven, and only the program part is what ties it to the concrete
implementation.

Additionally, View logic was slightly refactored with the inclusion of
a port-level defined view model. This provides an extra stage of decoupling
between the view and model implementations, hopefully reducing unnecessary
refactoring as the model develops.