Replace Gtk view with a Gdk based implementation
Gtk refused to comply with being put in a dock like UI display mode, being in
direct conflict with the desired functionality. Gdk complies with this excellently,
as well as being a lot less opinionated on being an application in its own right.

Gtk component usages are replaced by a custom UI library sitting on top of Gdk
and Cairo. The ui supports only very basic layout components, but it is in feature
parity with the removed Gdk view.