Implement shortcut fuzzy matching for modifiers
If an exact match for a keypress is not wound, we check if a match
can be found when ignoring the modifiers. Modifier downmatching only
extends to plain, modifierless shortcuts; S-C-x will not match with
S-x or C-x. This lets us defer solving the matching ambiguity for
later if or when such advanced downmatching becomes a desired feature.