Fix ShellExecutor async command related regression
After the introduction of error handling, specific programs failed
to launch. Cause was found to relate to the capture of stdio streams
regardless of whether the stream would be used or not.

The issue was circumvented by only piping the stdio streams with
synchronously executed commands. Relatedly, a small refactoring
of the ShellExecutor was made, enabling better error reporting even
when the program produces no stderr output.