Remove application-specific gui awareness from the gdk module
Building up on the previous modularization, the idea here is to make the Gdk
moudle less of a problem to deal with. Since all Gdk operations must happen
within the Gdk main thread, it was not possible to invert the dependency between
the Gdk and Gui modules in a simple way.

The already existing Gdk-thread channel messaging system is changed from passing
Gui specific datatypes to passing whole closures between threads. This, combined with
Gui side logic refactoring, enables making the Gdk side completely independent of
all Gui components and views, which now works through an enhanced Gui Renderer trait
instead.

The Gdk handle abstraction is expanded with the ability to store Gdk allocated
objects, enabling state to be transferred between multiple calls. The resulting
usage is a bit janky to use but it works, and the new Gui adapter is implemented
with it.