Prefer the term "mode" over "program" in longcut-core
Mode is a better name, because it conveys both independence and change, both of
which are things that the modes do. Additionally, it makes sense to speak about
how the program is currently in some mode, instead of how we're currently
running some specific program. The program we run is longcut, and longcut has
different modes.

Also add minor docstrings to all mode structs, and unify the naming around
error and contexts objects.