Support command output returning in Executor run_to_completion
This change enables handling command output in core, removing another obstacle
in the way of generating choose options using a command.

In the process the ShellModule was split apart into a Module and a Service,
substantially improving the testability. The module now contains the part
which integrates into the greater system, with the service containing the
actual functionality related to shell operations.