Support generating choose parameter options with a command call
It all finally comes together. The ChooseParameter is extended with optional
parameter generation command, which is used in the ParameterInputMode to call
out to the Executor in attempt to generate a list of options. The command
output is then split by the user-specifiable token, and made available as
options to pick from.