Clean up longcut-core module structure
The idea is to standardise the project structure on Module-Services-Config
triplets. Modules provide dependency injection entrypoints and are responsible
for instantiating the services within the modules and loading the module
configuration; the services expose the functionality of the module; and the
config is the common configuration shared between the services within the, but
which is only directly used within the services.