Refactor adapter code into separate crates
This enforces the conceptual separation between adapters joining two subdomains
together by preventing the two subdomains from accidentally getting coupled
together on the dependency level.

Note that this is not a universal separation; the longcut-gui package is so
tightly coupled to the longcut-core that even though the view adapter is moved
outside the package, the gui package itself still remains dependent on longcut-
core.

In process, some cleanup is done to lower the number of internal modules that
are exported by the entrypoints of different crates.