Add experimental Nix flake as an installation option
This makes installing Longcut much more convenient if you can access the Nix
package ecosystem, as the flake declaration handles the required dependencies
for you.

However, this is still experimental as I have no idea how the flake's dependencies
play together with other flakes that also depend on same libraries but in different
versions. Also, it would be great to idea to test if the flake works properly on a
real system running the NixOS operating system.